+++ src/main/java/com/talentotech/final_ecommerce/dto/CategoryDTO.java
@@ -1,6 +1,6 @@
 package com.talentotech.final_ecommerce.dto;
 
 public record CategoryDTO(
-        int categoria_id,
+        Long categoria_id,
         String nombre_categoria) {
 }
