+++ src/main/java/com/talentotech/final_ecommerce/model/Category.java
@@ -11,7 +11,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "categoria_id")
-    int categoriaId;
+    Long categoriaId;
 
     @Column(name="nombre_categoria")
     String nombreCategoria;
